requirements.txt
torch
torchvision
numpy
matplotlib
kagglehub
Pillow
